Finish utils refactor?
