+++ src/app/utils/airtable.ts
@@ -1,48 +1,9 @@
 'use server'
 
-import { getSession } from './auth-exposed'
+import { getSession } from './auth'
 import { person, updateShowInLeaderboard } from './data'
-import { getSelfPerson } from './server-only'
+import { getSelfPerson } from './server/airtable'
 
-export async function getPersonByMagicToken(token: string): Promise<{
-  id: string
-  email: string
-  slackId: string
-} | null> {
-  const baseId = process.env.BASE_ID
-  const apiKey = process.env.AIRTABLE_API_KEY
-  const table = 'people'
-
-  const url = `https://middleman.hackclub.com/airtable/v0/${baseId}/${table}?filterByFormula={magic_auth_token}='${encodeURIComponent(token)}'`
-
-  const response = await fetch(url, {
-    headers: {
-      Authorization: `Bearer ${apiKey}`,
-      'Content-Type': 'application/json',
-      'User-Agent': 'highseas.hackclub.com (getPersonByMagicToken)',
-    },
-  })
-
-  if (!response.ok) {
-    const err = new Error(`Airtable API error: ${await response.text()}`)
-    console.error(err)
-    throw err
-  }
-
-  const data = await response.json()
-  if (!data.records || data.records.length === 0) return null
-
-  const id = data.records[0].id
-  const email = data.records[0].fields.email
-  const slackId = data.records[0].fields.slack_id
-
-  if (!id || !email || !slackId) return null
-
-  return { id, email, slackId }
-}
-
-//FRONTEND
-// deprecate
 export async function getVotesRemainingForNextPendingShip(slackId: string) {
   const person = await getSelfPerson(slackId)
   return person['fields']['votes_remaining_for_next_pending_ship'] as number
@@ -62,8 +23,6 @@ export interface SafePerson {
   referralLink: string
 }
 
-// FRONTEND
-// Good method
 export async function safePerson(): Promise<SafePerson> {
   const record = await person()
