+++ src/app/utils/server/airtable.ts
@@ -39,7 +39,7 @@ export async function getPersonByAuto(num: string): Promise<{
     headers: {
       Authorization: `Bearer ${apiKey}`,
       'Content-Type': 'application/json',
-      'User-Agent': 'highseas.hackclub.com (getPersonByMagicToken)',
+      'User-Agent': 'highseas.hackclub.com (getPersonByAuto)',
     },
   })
 
@@ -61,7 +61,39 @@ export async function getPersonByAuto(num: string): Promise<{
   return { slackId }
 }
 
-export async function getSelfPersonIdentifier(slackId: string) {
-  const person = await getSelfPerson(slackId)
-  return person.fields.identifier
+export async function getPersonByMagicToken(token: string): Promise<{
+  id: string
+  email: string
+  slackId: string
+} | null> {
+  const baseId = process.env.BASE_ID
+  const apiKey = process.env.AIRTABLE_API_KEY
+  const table = 'people'
+
+  const url = `https://middleman.hackclub.com/airtable/v0/${baseId}/${table}?filterByFormula={magic_auth_token}='${encodeURIComponent(token)}'`
+
+  const response = await fetch(url, {
+    headers: {
+      Authorization: `Bearer ${apiKey}`,
+      'Content-Type': 'application/json',
+      'User-Agent': 'highseas.hackclub.com (getPersonByMagicToken)',
+    },
+  })
+
+  if (!response.ok) {
+    const err = new Error(`Airtable API error: ${await response.text()}`)
+    console.error(err)
+    throw err
+  }
+
+  const data = await response.json()
+  if (!data.records || data.records.length === 0) return null
+
+  const id = data.records[0].id
+  const email = data.records[0].fields.email
+  const slackId = data.records[0].fields.slack_id
+
+  if (!id || !email || !slackId) return null
+
+  return { id, email, slackId }
 }
